requirements.txt
streamlit
<<<<<<< HEAD
transformers
torch
accelerate
python-dotenv
sentencepiece
=======
groq
python-dotenv
>>>>>>> 8ea3ce0 (Initial commit)
